requirements.txt
requests
beautifulsoup4
streamlit
unidecode
googlesearch-python
pandas
folium
streamlit-folium